Working queue slice insertion implementation!
Fixed most of the previous issues by inserting _after_ the existing
event slice. This does a number of things:
- Lets us atomicly mark the previous slice top as removed
- Prevents any in-flight changes from affecting the previous slice top
  (assuming CAS succeeds)
- Simplifies conditions for dispatch, assuming dispatch fixes all
  removes up to the target timestamp before dispatching events

This just works out really nicely, requiring only two nuanced, but
simple loops.

A funny side-effect is that we don't actually need to finish removals
during insertion, because we can expect later traversals to fix the
queue for us. The temporary state consumes no extra memory, is
guaranteed to be fixed by dispatch, and adds at most 1 extra event
node to the queue for each slice.