Added a, somewhat taped together, delay hammer example
This actually showcases the pending slices, though it requires
reducing the granularity to be noticable at human perception.

Though that does show that the insertion cost for a uniform distribution
of delays is pretty suredly O(n) as long as your delays are decently larger
than the granularity of the timer.