Attempted a different scheme that was more promising
But so far have not found a solution that works. At this point I think
it may be best to accept defeat with regards to a complete non-blocking
implementation, at least for now.

Even if a non-blocking implementation is possible, currently it looks like
cancel would end up being worst case O(n) best case O(sqrt(n)), which, even
when non-blocking, is inferior to the O(1) cancel of the previous design...