Reimplemented post using locking primitives, also O(1) dispatch startup
Since we already need back-references everywhere for an efficient
cancel, we can add a back-reference for siblings to avoid reversing
the linked-list of dequeued siblings.

It's a minor improvement, since you still need to iterate all of those
entries eventually, but it does:
- decrease the time to start execute events
- avoids messy locking since siblings could be deleted at any time
- avoids invalidating slice traversals as much as possible since we
  don't invalidate the head of a slice every push anymore
- simplify the code a bit

I originally avoided this because of the extra RAM cost for each event,
since it would push the event head from a very nicely aligned 8 words to
a terrible 9 words in a structure that would be _very_ nice to have as
little overhead for as possible. But I realized we don't need to mark
every eptr, since siblings and back-references are only ever read while
locked. If we cram two back-references in a word this makes the
resulting event header only 7 words, which is great! We can fit closures
with up to a single word of storage (very common) into the smallest
power-of-two event.

Also switched to a fixed encoding size, this is less flexible, but
should result in much better code generation, it also means you need to
passing around Equeue references in less places.

uatom == u64: [- 8-|- 8-|-- 16 --]
uatom == u32: [-16-|-16-|-- 32 --]
                 ^    ^      ^-- compressed eptr
                 |    '--------- generation count of eptr
                 '-------------- mark verifying this marked-eptr