Moved periodic event cancel check into the nearest lock: enqueue
This is a bit messy, but seems to be the best solution. Now canceling
periodic events should always be successful in the sense that a periodic
event will not be placed back into the queue after canceled is called

Previously there was a (still safe) race condition where a canceled
periodic event could be enqueued if it didn't see the cancelation before
calling enqueue.