Found a method for non-locking linked-list traversal without mark+gen pairs
As long as we check the previous next pointer before "trusting" the current
next pointer, we should be able to safely make progress without the
mark+gen pairs.

We still need a generation count to prevent ABA problems, but that does
at least open up a 1-byte field for other purposes (or larger
pointers/generation counts).