Implemented trigger and fundamental async functions
Equeue::trigger is a binary-semaphore-esque alternative to
post for static events. It's very useful for implementing the wake
functions required for Futures.

Also implemented async yield/sleep/timeout. Unfortunately, despite
the infrastructure available to do so, it's not possible to avoid the
additional allocation of events for sleep/timeouts. This is entirely due
to how limited the Context/Waker API is, which we already have to work
around in order to run Futures at all on non-global schedulers without
dynamic memory.