Fixed leaked static events
If not pending when the Equeue is dropped, we weren't cleaning up
events, since they weren't in our pending list. Aside from the obvious
memory leak, this also lead to late wakeups causes segfaults by trying
to interact with the Equeue after being torn down.

The solution is to instead search for any non-dropped events in our
slab. This is more expensive, but you can't really expect Equeue drop
to be a cheap operation.

If anything it's fortunate that segfaults were being triggered, if this
just caused a memory leak it may not have been noticed for a while.
Especially since it only happens when destroying Equeues, which is fairly
uncommon.