Adopted a different, more controlled, approach to event dequeueing
In the original design, event dequeueing worked by relying on a global,
monotonically increasing, tick and generation count. Whenever we dispatch,
we update to the new tick+generation count, implicitly marking all
events as dispatching. The separate generation count helps prevent
infinite loops by incrementing before new events can be recursively
enqueued.

This works, but:

1. It requires atomic/locked access to the tick + generation count. Which may
   be difficult if ticks are a large size, such as 64-bits on a 32-bit system.

2. Since all ready events are marked as dispatching simultaneously, it
   could actually take us a while to get to any individual event,
   preventing memory reclaimation during cancel in the meantime.

In the new design, the first apporach was to not use the global tick
for synchronization, but instead keep the dequeued events in a state
where cancel would still work.

But this became a bit tricky. Since we are now using compressed
pointers, we can't just reference the on-stack pointer that would be
used to temporarily hold the dispatching events. So initially I tried
marking the first event as dispatching, even though it has to wait
for the _actual_ dispatching event to execute.

This also worked, though it meant you couldn't reclaim the next event,
which is a bit weird.

But then I tried to change the unqueueing in dispatch to use unqueue,
which would consolidate all racey queue operation into three strongly
intentioned functions. This made the unqueue implementation complicated,
with a number of special corner cases to handle.

So back to the drawing board.

The new design here uses a second global "dequeue" pointer to hold
events ready to execute for a given dispatch call. Unqueueing an event
to dispatch goes through the same "unqueue" operation that cancel does,
allowing the code to be consolidated quite nicely.

This makes the dispatch implementation extremely well defined, though
does mean a bit more logic to handle updating the queue and dequeue
pointers. It even means parallel dispatch calls should properly
load-balance (probably, I haven't tested this).