Adopted mutex guards
I originally tried using a callback based lock API because it enforced
brackets that tightly bind the point of synchronization, and because it
allows a safe, trait-compatible API.

However being able to return/break/continue from synchronized scopes is
just too useful, especially for the lockless algorithms in play here,
which often require loops with many continue statements for various
corner cases.

Downside, the implementation needs to be unsafe since we can't use
lifetimes in traits. But I think the tradeoff is worth it to avoid
the Option/Result/Either abuse that was required.