Fixed ordering issue caused by a race condition between dequeue + enqueue
When we dequeue, we take a set of slices, and make them ready:

    queue  now
    v      v
    '':'':::''''::''':'':'':::::::'''
         : '    :    :  :  ' :  ''
         :

Becomes:

    dequeue  queue
    v        v
    '':''::  :''''::''':'':'':::::::'''
         :   '    :    :  :  ' :  ''
         :

But if we only update the last event in the dequeue list to mark it
null, the other events remains untouched.

I overlooked this deviation from Harris's lock-free linked-list,
consider what happens if we dequeue while some thread is enqueueing:

       enqueue
       |
    queue  now
    v  v   v
    '':'':::''''::''':'':'':::::::'''
         : '    :    :  :  ' :  ''
         :

Becomes:

       enqueue
       |
    dequeue  queue
    v  v     v
    '':''::  :''''::''':'':'':::::::'''
         :   '    :    :  :  ' :  ''
         :

Normally, during cancels, we always update the event that is removed,
invalidating any in-flight enqueues. But here enqueue doesn't notice
the state of the queue has changed.

We should never end up losing an event, but consider what happens if
enqueue wants to enqueue somewhere later in the queue:

           enqueue
           |
    dequeue|  queue
    v      v  v
    '':''::'  :''''::''':'':'':::::::'''
         :    '    :    :  :  ' :  ''
         :

We end up inserting a new event for _immediate_ dispatch, even if it
should be enqueued much later.

One, albiet a bit expensive, fix, is to check for dequeue changes during
the enqueue traversal. This causes enqueue to restart in the above
situation.

I say it's expensive, because it unfortunately causes all enqueues
to be restarted on every dequeue operation.