Found a slightly better enqueue+dequeue strategy
This slight tweak checks the current queue head target
before fully invalidating an enqueue operation when dequeue happens.

This means dequeue won't always invalidate all current enqueue operations,
giving enqueue a better chance to succeed. There's still a chance for a
false negative, if we are enqueueing on the first slice (which is why we
can't rely on the head target check alone), but fortunately in that case
the extra search should be only O(1) instead of O(n).