Reworked atomic definitions, removed AtomicUsize, now with feature flags
Removed AtomicUsize, so now the only dependency is on AtomicDeptr and
AtomicEptr (though we probably don't need the AtomicEptr for lockless
implementations).

Also added some flags to control which integer sizes to use.
Unfortunately because of the lack of primitive integer traits in the
Rust stdlib, it's pretty tedius to use generics for configurable integer
widths. However leaving this up to build time configuration is pretty
reasonable.