Refactored wait_async implementations a bit, and fixed multi-consumer issues
This is a perfect example of feature creep.

Equeue is intended to provide a synchronization facility for asynchronous
code (either threads or interrupts depending on your system). So you would
expect Equeue to be marked as Send+Sync, marker traits in Rust which indicate
your type is thread-safe.

This is all fine and good, except in Rust, Send+Sync only applies to
whole objects, not individual functions. This means if Equeue is marked
as Send+Sync, _all_ methods must be synchronized. Originally intended
as a multi-producer, single-consumer sort of system, this raises the
question, what happens if you call dispatch from multiple threads?

Fortunately, due to the synchronization required between producers
and consumers, synchronizing dispatch actually comes for free. And with
a bit of tweaking it can actually be well behaved, distributing events
across whichever threads are dispatching. Well, originally the first thread
that arrived would get all ready events, potentially starving other
consumers, but it was just a bit of tweaking which also helped simplify
the code (deduplicated cancel and unqueue).

So here we are, with the _possibility_ that a user calls dispatch from
multiple threads, and Equeue potentially distributing events across
threads evenly. Except this suddenly changes the expected behavior and
implementation of the underlying semaphore. Originally it was a simple
binary semaphore, implementable with an atomic bool that was cleared
by which ever thread woke up. But we need to wake up _every_ thread,
otherwise only on thread will do all the work.

For systems with multi-cores, this means that the underlying semaphore
implementation should use some primitive, such as a conditional
semaphore, in order to wake up all events. Annoyingly, this is more
complicated, and potentially tricky to get right.

All for a library that is most likely used only single-core systems.

Fortunately, those can use the original binary semaphore, and ignore
the unreasonable possiblity of the multiple dispatches on multiple
threads. They would only end up starving each other anyways.