Removed most raw operations, leaving alloc/dealloc_raw and Event::from/into_raw
This is a simpler/easier-to-maintain alternative to hiding raw events behind
feature flags. We still keep the unsafe raw allocations necessary for
dynamically sized events, but expose the minimum API to get from a
raw event to the safe Event type.

This also makes it possible to allocate unsized types, which now permeat
the Event implementation. This required a bit of moving around the
pointers inside Event in order to store the unsized-type's fat pointer
correctly.