Implemented a functioning lock-free queue implementation using a helper FSM
I initially rejected a completely lock-free implementation due to the
complexity of the queue implementation. It's not just a linked-list, which has
working solutions such as Harris's linked-list, but a linked-list of
linked-lists, which turns out to create a lot of problems. Most notably
when the node that is shared between both linked-lists needs to be
removed.

However, after reading up a bit more about the space, most notably
Greenwald's "Two-Handed Emulation: How to build Non-Blocking
implementations of Complex Data-Structures using DCAS" paper (note DCAS is an
instruction not actually available in most hardware), I think I've found a
workable solution using only traditional CAS operations with no staging or
extra control blocks. Though this is mostly possible because equeue only needs
the context available in the existing event structs to complete every necessary
operation.

---

The basic idea is instead of using locks/mutexes to synchronize
mutation, synchronize by using a shared state machine that any thread
can help progress. To keep these "helper" threads in sync, all of the
moving parts of the state machine are stored in CAS words with embedded
counters that must be kept in lockstep.

For equeue, the first and most important word contains the actual
state machine and enough context to complete the operation. Fortunately
for us we only need a compressed eptr to a prepared ebuf struct:

    help_op: [- gen -|-state-|--   eptr   --]

Each state in our state machine is a single CAS of some part of the
queue's data structure.

This single word has all the info necessary to perform the operation,
but if we let every helper thread run with this we would end up with
garbage as each successful CAS puts the queue into a different state!

We need some way to get all of the helper threads to agree on the
context of the CAS. This is accomplished by a few more CAS words of
state, each with a generation count with one important rule: The
generation count must match the help_op's generation count!

    help_ctx: [- gen -|-------|--   eptr   --]
    help_old: [- gen -|---   old value    ---]

If the generation count is outdated by exactly 1, we can calculated
the required value and update it. If it already matches the help_op's
generation count, someone beat us to updating it. And if it is any other value,
any number of operations have occured so we need to restart.

Once both the help_ctx and help_old values have been decided, we can
perform the in-queue CAS. It's up to the actual operation to avoid
normal CAS issues such as ABA, but once the CAS occurs, any other helper
threads should, worst case, fail.

In effect, this forces helper threads to reach an agreement before
executing a CAS. It's important that both the destination and previous
value are agreed upon, otherwise different helper threads can end up
trying different CASs due to ongoing mutation.

---

This has been the biggest piece of work in equeue recently, requiring a
transformation of the core algorithm into a state machine with limited
auxilary state. But it seems to be working!