Fixed dequeue/queue mark issue with invalidating traversals
This is the reoccurence of a tricky to test bug that was fixed
in 26aa00a.

The source fo the problem is that we relied on the incrementing of the mark
in dequeue to invalidate enqueue traversals, but we load queue to start
the traversal. By itself this isn't an issue, but makes the algorithm
very sensitive to update order. Specifically we need to update the queue
pointer before we update the dequeue pointer.

This is a bit of an issue for our helper FSM. Since we're limited to
very little state, updating the queue first means we lose the pointer
we need to be storing into dequeue. Updating the dequeue first avoids
this but reintroduces the ordering race condition from 26aa00a, since
we could begin an enqueue traversal with the new dequeue pointer and
old queue pointer. This does not get caught by our helper FSM logic
since we only check the updated dequeue pointer before committing
an enqueue.

The solution adopted in this commit is to simply have two separate
marks, one for dequeue and one for queue. These both get updated when
we perform a dequeue operation, but ensures we can't load an old queue
pointer without also getting the old mark, which would be invalidated
by the time the helper FSM logic allows us to commit the enqueue
operation.