Added exhaustive testing via loom for some limited test cases
The Loom crate is quite powerful, and an invaluable resource for testing
lock-less programs, perfect for increasing confidence in equeue's
implementation.

Note that due to state-space explosion, the test cases under loom are
very small, enqueueing a whopping 3 events from 2 threads. But, since we
are dealing with exhaustive testing, we don't need a large input to find
bugs.