Simplified the helper FSM by moving context into single atomic integer
This sort of fell out of the observation that we can still synchronize
the multipurpose mark field without storing it in our helper context
explicitly.

If we only need the location and generation count of the old value
being CASed, we can fit this into a single atomic integer thanks
to our compressed eptrs:

    [-fsm gen-|-old gen-|--    eptr   --]

This does rely entirely on the generation count for enforcing uniqueness
of the values we are CASing, but if we can't rely on this the CASs under
the FSMs are broken anyways.