Added a very rough code-size benchmark
This was surprisingly tricky to measure, Rust does not have great
tooling yet for measuring code size of arbitrary libraries.

Cargo bloat is a great tool, but only works at the binary level. This
require a bit of a funky benchmark whose sole purpose is to link
in the relavent code to measure.

Even then I wasn't able to separate many of the debug related
functionality such as panic and Debug impls, which may be brought
in for debugger support, not sure.

It's not a perfect measurement, and the functions measured are somewhat
arbitrary, but should be sufficient for comparison between implementations.
(What I'm saying is this has no value outside of relative measurements).