Added filtering out of Debug impls for code size measurements
I have no idea how these are still getting dragged in, since the
binary needs to be executable for cargo bloat to work, it may be
due to _start setup, some hidden backtrace path, or maybe as a
hook for debuggers. But either way it drags in a theoretically
unneccessary 1KiB of extra code, which accounts for ~12% of the
entire codebase.

Filtering out symbols with the string Debug is a working, if
heavy-handed and inaccurate, solution which is the best I have
for now.