Added benchmarking for operations and lock hold times
The overall runtimes were perfectly suited for Criterion, measuring
the lock hold times however required a bit more thinking, since
intrusively measuring the time a lock holds took more time than the
actual functions being measured.

The solution here is to use a separate thread to sample a flag
indicating whether or not the lock is held.

Setting the flag had neglible impact on the overall runtime, while the
sampling allow us to still build up an estimate.

The result may be a bit innaccurate, but over repeated tests seem pretty
consistent. This is probably thanks to how Criterion run several million
samples under the same measurement.

---

A bit confusing, but Criterion seemed unhappy with zero runtime
measurements, which are the expected value from lockless-lock-times.
I'm not sure why this is, since zero seems like a perfectly reasonable
result for custom measurements, but for now I've just bumped the minimum
to 1 nanosecond. This doesn't seem to affect the results.